Add Module & Code Refactor
Add js modules in modules dir
Add ES6 syntax
