+++ index.html
@@ -4,8 +4,8 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Awesome Books</title>
-    <link rel="stylesheet" href="styles/reset.css" type="text/css" media="all"/>
-    <link rel="stylesheet" href="styles/style.css" type="text/css" media="all"/>
+    <link rel="stylesheet" href="./assets/cs/reset.css" type="text/css" media="all"/>
+    <link rel="stylesheet" href="./index.css" type="text/css" media="all"/>
   </head>
   <body>
     <header class="nav-cont">
@@ -68,7 +68,7 @@
     <footer class="footer">
       Copyright...
     </footer>
-    <script src="scripts/awesome_mod.js"></script>
-    
+    <script type="module" src="./index.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/luxon@2.3.2/build/global/luxon.min.js"></script> 
   </body>
 </html>
