+++ modules/clearinputs.js
@@ -1,0 +1,8 @@
+import { titleInput, authorInput } from './navigation.js';
+
+const resetInput = () => {
+  titleInput.value = '';
+  authorInput.value = '';
+};
+
+export default resetInput;
