+++ modules/navigation.js
@@ -1,0 +1,18 @@
+const menuLinks = document.querySelectorAll('.link-item');
+
+const date = document.querySelector('.date-item');
+
+const welcomeMsg = document.querySelector('.welcome-msg');
+
+const libraryContainer = document.querySelector('#library-container');
+
+const form = document.querySelector('.form-container');
+const titleInput = document.querySelector('#title');
+const authorInput = document.querySelector('#author');
+const submitBtn = document.querySelector('#btn');
+
+const contact = document.querySelector('.contact');
+
+export {
+  menuLinks, date, welcomeMsg, libraryContainer, form, titleInput, authorInput, contact, submitBtn,
+};
